back-end/requirements.txt
h5py==2.10.0
Flask==1.0.2
flask_restplus==0.12.1
keras==2.2.4
tensorflow==1.14.0
werkzeug==0.16.0
numpy==1.17.2
scipy==1.5.2
scikit-learn==0.23.2
Pillow==7.2.0
gunicorn==20.0.4
boto3==1.16.9
pymongo==3.11.0
python-dotenv==0.15.0
